+++ framework_src/freswift_native_extension/ane/build_osx.sh
@@ -9,7 +9,7 @@ echo $pathtome
 PROJECT_NAME=FreSwiftExampleANE
 FRESWIFT_NAME=FreSwift
 
-AIR_SDK="/Users/eoinlandy/SDKs/AIRSDK_31"
+AIR_SDK="/Users/eoinlandy/SDKs/AIRSDK_32"
 
 #Setup the directory.
 echo "Making directories."